The mash() loop now waits for data instead of a timeout
This way the mash() doesn't press any buttons until you tell it to by
sending a character over serial, so in case the typewriter needs to be
reset, it can be done without having to race the timeout.

Also, I made all the comments into the same style and fixed some typos
that were preventing a compile.