Faster caps handling
When printing all-caps things, the performance used to be 3x worse, as
for every capital letter the typewriter would have to do 3 keystrokes.
Now, the Arduino pays attention to the state of the caps lock, and only
changes it when it needs to be changed. This is more human-like and
makes things go so much faster, especially when printing ASCII art (see
bieber.txt). I haven't found a solution for handling repeats of special
characters, so those remain slow (except for some hard-coded
enhancements I made to # and :). Capital letters are mathematically
offset from their lowercase bretheren, and special characters are kinda
arbitrarily laid out on the keyboard, so there was no obvious magic I
could do.

I also moved handling of space, newline, and carriage returns to the
main switch for faster and more direct printing.