Completely Rewrite Tweetmash
Now it actually formats the tweets before they are printed. The file is
parsed into a list of Tweet objects, each having an author and a
message. Then, one-by-one, the list of tweets is printed, making sure
that each word in its entirety will fit on the line (real, bona-fide
word wrap). It also arranges the page into two columns: twitter handle
on the left, and tweets on the right (all justified to 20 characters
from the left margin).

This isn't very DRY programatically, but it works.